Move tileToQuadkey to tests
