+++ test/modules/carto/layers/quadbin-tile-layer.spec.js
@@ -3,13 +3,41 @@ import {generateLayerTests, testLayerAsync} from '@deck.gl/test-utils';
 import {_QuadbinTileLayer as QuadbinTileLayer} from '@deck.gl/carto';
 import {WebMercatorViewport} from '@deck.gl/core';
 import {testPickingLayer} from '../../layers/test-picking-layer';
-import {tileToQuadbin, tileToQuadkey, quadbinToTile} from '@deck.gl/carto/layers/quadbin-utils';
+import {tileToQuadbin, quadbinToTile} from '@deck.gl/carto/layers/quadbin-utils';
 
 const TILES = ['https://tile-url.com/{i}'];
 const TILEJSON = {
   maxresolution: 10,
   tiles: TILES
 };
+
+export function tileToQuadkey(tile) {
+  let index = '';
+  for (let z = tile.z; z > 0; z--) {
+    let b = 0;
+    const mask = 1 << (z - 1);
+    if ((tile.x & mask) !== 0) b++;
+    if ((tile.y & mask) !== 0) b += 2;
+    index += b.toString();
+  }
+  return index;
+}
+
+function quadkeyToTile(quadkey) {
+  const tile = {x: 0, y: 0, z: quadkey.length};
+
+  for (let i = tile.z; i > 0; i--) {
+    const mask = 1 << (i - 1);
+    const q = Number(quadkey[tile.z - i]);
+    if (q === 1) tile.x |= mask;
+    if (q === 2) tile.y |= mask;
+    if (q === 3) {
+      tile.x |= mask;
+      tile.y |= mask;
+    }
+  }
+  return tile;
+}
 
 test('QuadbinTileLayer', async t => {
   const testCases = generateLayerTests({
@@ -42,22 +70,6 @@ test('QuadbinTileLayer tilejson', async t => {
   await testLayerAsync({Layer: QuadbinTileLayer, testCases, onError: t.notOk});
   t.end();
 });
-
-function quadkeyToTile(quadkey) {
-  const tile = {x: 0, y: 0, z: quadkey.length};
-
-  for (let i = tile.z; i > 0; i--) {
-    const mask = 1 << (i - 1);
-    const q = Number(quadkey[tile.z - i]);
-    if (q === 1) tile.x |= mask;
-    if (q === 2) tile.y |= mask;
-    if (q === 3) {
-      tile.x |= mask;
-      tile.y |= mask;
-    }
-  }
-  return tile;
-}
 
 test('QuadbinTileLayer autoHighlight', async t => {
   await testPickingLayer({
